requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
av==14.1.0
beautifulsoup4==4.13.3
bertopic==0.17.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
databricks-sdk==0.43.0
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
distlib==0.3.9
docker==7.1.0
easyocr==1.7.2
executing==2.2.0
fastapi==0.115.12
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
lmdb==1.6.2
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mlflow==2.20.1
mlflow-skinny==2.20.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
git+https://github.com/JiahuiYu/neuralgym.git
ninja==1.11.1.3
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyngrok==7.2.4
pynndescent==0.5.13
pyparsing==3.2.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tokenizers==0.21.0
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
umap==0.1.1
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
